+++ packages/js-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mnfst/sdk
 
+## 1.3.0
+
+### Minor Changes
+
+- beb4317: Replaced incremental int IDs by UUIDs, thanks @jerryjappinen
+
+## 1.2.2
+
+### Patch Changes
+
+- da62deb: Fixed PATCH request password update issue and return token on JS SDK
+
 ## 1.2.1
 
 ### Patch Changes